Removed a large number of warnings from iOS library build.

+++ include/align.h
@@ -15,7 +15,7 @@ typedef struct align { /* alignment & record */
 } align;
 
 /* bounds for "record" -- respect initial alignments of 10 */
-#define ALIGNLIM (10L + (moves / 200L))
+#define ALIGNLIM ((int)(10L + (moves / 200L)))
 
 #define A_NONE (-128) /* the value range of type */
